@@ -1,14 +1,6 @@
 package com.baeldung.reactive.debugging.consumer;
 
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import ch.qos.logback.classic.spi.IThrowableProxy;
-import com.baeldung.reactive.debugging.consumer.model.Foo;
-import com.baeldung.reactive.debugging.consumer.service.FooService;
-import com.baeldung.reactive.debugging.consumer.utils.ListAppender;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Hooks;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -16,7 +8,16 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import com.baeldung.reactive.debugging.consumer.model.Foo;
+import com.baeldung.reactive.debugging.consumer.service.FooService;
+import com.baeldung.reactive.debugging.consumer.utils.ListAppender;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import ch.qos.logback.classic.spi.ILoggingEvent;
+import ch.qos.logback.classic.spi.IThrowableProxy;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Hooks;
 
 public class ConsumerFooServiceIntegrationTest {
 
@@ -37,26 +38,26 @@ public class ConsumerFooServiceIntegrationTest {
         service.processFoo(flux);
 
         Collection<String> allLoggedEntries = ListAppender.getEvents()
-            .stream()
-            .map(ILoggingEvent::getFormattedMessage)
-            .collect(Collectors.toList());
+                .stream()
+                .map(ILoggingEvent::getFormattedMessage)
+                .collect(Collectors.toList());
 
         Collection<String> allSuppressedEntries = ListAppender.getEvents()
-            .stream()
-            .map(ILoggingEvent::getThrowableProxy)
-            .flatMap(t -> {
-                return Optional.ofNullable(t)
-                    .map(IThrowableProxy::getSuppressed)
-                    .map(Arrays::stream)
-                    .orElse(Stream.empty());
-            })
-            .map(IThrowableProxy::getClassName)
-            .collect(Collectors.toList());
+                .stream()
+                .map(ILoggingEvent::getThrowableProxy)
+                .flatMap(t -> {
+                    return Optional.ofNullable(t)
+                            .map(IThrowableProxy::getSuppressed)
+                            .map(Arrays::stream)
+                            .orElse(Stream.empty());
+                })
+                .map(IThrowableProxy::getClassName)
+                .collect(Collectors.toList());
         assertThat(allLoggedEntries).anyMatch(entry -> entry.contains("The following error happened on processFoo method!"))
-            .anyMatch(entry -> entry.contains("| onSubscribe"))
-            .anyMatch(entry -> entry.contains("| cancel()"));
+                .anyMatch(entry -> entry.contains("| onSubscribe"))
+                .anyMatch(entry -> entry.contains("| cancel()"));
 
         assertThat(allSuppressedEntries)
-            .anyMatch(entry -> entry.contains("reactor.core.publisher.FluxOnAssembly$OnAssemblyException"));
+                .anyMatch(entry -> entry.contains("reactor.core.publisher.FluxOnAssembly$OnAssemblyException"));
     }
 }
