@@ -19,50 +19,50 @@ import com.baeldung.micronaut.vs.springboot.CompareApplication;
 @SpringBootTest(classes = CompareApplication.class)
 @AutoConfigureMockMvc
 public class ArithmeticControllerUnitTest {
-	@Autowired
+    @Autowired
     private MockMvc mockMvc;
-	
-	@Test
-	public void givenTwoNumbers_whenAdd_thenCorrectAnswerReturned() throws Exception {
-		Float expected = Float.valueOf(10 + 20);
-		this.mockMvc.perform(MockMvcRequestBuilders.get("/math/sum/10/20")
-				.accept(MediaType.APPLICATION_JSON))
-	            .andExpect(status().isOk())
-	            .andExpect(content().string(expected.toString()));
-	}
-	
-	@Test
-	public void givenTwoNumbers_whenSubtract_thenCorrectAnswerReturned() throws Exception {
-		Float expected = Float.valueOf(20 - 10);
-		this.mockMvc.perform(MockMvcRequestBuilders.get("/math/subtract/20/10")
-				.accept(MediaType.APPLICATION_JSON))
-	            .andExpect(status().isOk())
-	            .andExpect(content().string(expected.toString()));
-	}
 
-	@Test
-	public void givenTwoNumbers_whenMultiply_thenCorrectAnswerReturned() throws Exception {
-		Float expected = Float.valueOf(20 * 10);
-		this.mockMvc.perform(MockMvcRequestBuilders.get("/math/multiply/20/10")
-				.accept(MediaType.APPLICATION_JSON))
-	            .andExpect(status().isOk())
-	            .andExpect(content().string(expected.toString()));
-	}
-	
-	@Test
-	public void givenTwoNumbers_whenDivide_thenCorrectAnswerReturned() throws Exception {
-		Float expected = Float.valueOf(20 / 10);
-		this.mockMvc.perform(MockMvcRequestBuilders.get("/math/divide/20/10")
-				.accept(MediaType.APPLICATION_JSON))
-	            .andExpect(status().isOk())
-	            .andExpect(content().string(expected.toString()));
-	}
-	
-	@Test
-	public void whenMemory_thenMemoryStringReturned() throws Exception {
-		this.mockMvc.perform(MockMvcRequestBuilders.get("/math/memory")
-		    .accept(MediaType.APPLICATION_JSON))
-		    .andExpect(status().isOk())
-		    .andExpect(content().string(containsString("Initial:")));
-	}
+    @Test
+    public void givenTwoNumbers_whenAdd_thenCorrectAnswerReturned() throws Exception {
+        Float expected = Float.valueOf(10 + 20);
+        this.mockMvc.perform(MockMvcRequestBuilders.get("/math/sum/10/20")
+            .accept(MediaType.APPLICATION_JSON))
+            .andExpect(status().isOk())
+            .andExpect(content().string(expected.toString()));
+    }
+
+    @Test
+    public void givenTwoNumbers_whenSubtract_thenCorrectAnswerReturned() throws Exception {
+        Float expected = Float.valueOf(20 - 10);
+        this.mockMvc.perform(MockMvcRequestBuilders.get("/math/subtract/20/10")
+            .accept(MediaType.APPLICATION_JSON))
+            .andExpect(status().isOk())
+            .andExpect(content().string(expected.toString()));
+    }
+
+    @Test
+    public void givenTwoNumbers_whenMultiply_thenCorrectAnswerReturned() throws Exception {
+        Float expected = Float.valueOf(20 * 10);
+        this.mockMvc.perform(MockMvcRequestBuilders.get("/math/multiply/20/10")
+            .accept(MediaType.APPLICATION_JSON))
+            .andExpect(status().isOk())
+            .andExpect(content().string(expected.toString()));
+    }
+
+    @Test
+    public void givenTwoNumbers_whenDivide_thenCorrectAnswerReturned() throws Exception {
+        Float expected = Float.valueOf(20 / 10);
+        this.mockMvc.perform(MockMvcRequestBuilders.get("/math/divide/20/10")
+            .accept(MediaType.APPLICATION_JSON))
+            .andExpect(status().isOk())
+            .andExpect(content().string(expected.toString()));
+    }
+
+    @Test
+    public void whenMemory_thenMemoryStringReturned() throws Exception {
+        this.mockMvc.perform(MockMvcRequestBuilders.get("/math/memory")
+            .accept(MediaType.APPLICATION_JSON))
+            .andExpect(status().isOk())
+            .andExpect(content().string(containsString("Initial:")));
+    }
 }
