@@ -8,8 +8,8 @@ import org.springframework.test.web.reactive.server.WebTestClient.ResponseSpec;
 
 /**
  * In order to run this live test, start the following classes:
- * - com.baeldung.debugging.server.ServerDebuggingApplication
- * - com.baeldung.debugging.consumer.ConsumerDebuggingApplication
+ * - com.baeldung.reactive.debugging.server.ServerDebuggingApplication
+ * - com.baeldung.reactive.debugging.consumer.ConsumerDebuggingApplication
  */
 public class ConsumerFooServiceLiveTest {
 
